+++ src/api/client/server.rs
@@ -146,6 +146,7 @@ mod test {
     use actix_http::{ws::Message, HttpService};
     use actix_http_test::{TestServer, TestServerRuntime};
     use futures::{future::IntoFuture as _, sink::Sink as _, Stream as _};
+    use medea_client_api_proto::{CloseDescription, CloseReason};
 
     use crate::{
         api::control, conf::Conf, signalling::Room,
@@ -230,9 +231,22 @@ mod test {
                                 read.into_future()
                                     .map_err(|(e, _)| panic!("{:?}", e))
                                     .map(|(item, _)| {
+                                        let description = CloseDescription::new(
+                                            CloseReason::Idle,
+                                        );
+                                        let close_reason = ws::CloseReason {
+                                            code: ws::CloseCode::Normal,
+                                            description: Some(
+                                                serde_json::to_string(
+                                                    &description,
+                                                )
+                                                .unwrap(),
+                                            ),
+                                        };
+
                                         assert_eq!(
                                             Some(ws::Frame::Close(Some(
-                                                ws::CloseCode::Normal.into()
+                                                close_reason
                                             ))),
                                             item
                                         );