+++ src/signalling/participants.rs
@@ -24,7 +24,7 @@ use futures::{
     Future,
 };
 
-use medea_client_api_proto::Event;
+use medea_client_api_proto::{CloseDescription, CloseReason, Event};
 
 use crate::{
     api::{
@@ -195,7 +195,11 @@ impl ParticipantService {
             {
                 ctx.cancel_future(handler);
             }
-            Box::new(wrap_future(connection.close().then(move |_| Ok(member))))
+            Box::new(wrap_future(
+                connection
+                    .close(CloseDescription::new(CloseReason::Reconnected))
+                    .then(move |_| Ok(member)),
+            ))
         } else {
             Box::new(
                 wrap_future(self.turn.create(
@@ -299,7 +303,10 @@ impl ParticipantService {
         let mut close_fut = self.connections.drain().fold(
             vec![],
             |mut futures, (_, mut connection)| {
-                futures.push(connection.close());
+                futures.push(
+                    connection
+                        .close(CloseDescription::new(CloseReason::Finished)),
+                );
                 futures
             },
         );