+++ proto/client-api/CHANGELOG.md
@@ -12,7 +12,8 @@ All user visible changes to this project will be documented in this file. This p
 ### Added
 
 - `TrackId` and `PeerId` types ([#28]);
-- `Incrementable` trait ([#28]).
+- `Incrementable` trait ([#28]);
+- `CloseReason` and `CloseDescription` types ([#58](/../../pull/58)).
 
 [#28]: /../../pull/28